requirements.txt
pymupdf>=1.24.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
ollama>=0.4.0
pandas>=2.0.0
numpy>=1.26.0
tqdm>=4.66.0
python-dotenv>=1.0.0
jupyter>=1.0.0
streamlit>=1.30.0
